\name{sl.curvilin2unstr}
\alias{sl.curvilin2unstr}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Convert Curvilinear to Unstructured
}
\description{
THIS FUNCTION IS NOT YET IMPLEMENTED BUT WILL BE NEEDED! (The current code is a stub; the function will exit immediately when called) +++ Convert a curviliear mesh, where fields are provided in matrices and lat-lon coordinates are provided either in matrices (general) or vectors (only for lat-lon meshes), into an identical unstructured mesh. The conversion allows to apply all the unstructured-mesh algorithms also to curvilinear-mesh data.
}
\usage{
sl.curvilin2unstr(vals = NULL, byrow = TRUE, convert.grid = FALSE, lon.c = NULL, lat.c = NULL, lon.i = NULL, lat.i = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{vals}{
To be described ...
}
  \item{byrow}{
To be described ...
}
  \item{convert.grid}{
To be described ...
}
  \item{lon.c}{
To be described ...
}
  \item{lat.c}{
To be described ...
}
  \item{lon.i}{
To be described ...
}
  \item{lat.i}{
To be described ...
}
}
\details{
To be described ...
}
\value{
To be described ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Helge Goessling
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
## To be provided ...
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
