\name{sl.polygon.rotdir}
\alias{sl.polygon.rotdir}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Check Polygon Direction of Rotation
}
\description{
UNFINISHED ### Check the direction of rotation of a polygon on a sphere.
}
\usage{
sl.polygon.rotdir(lon, lat)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{lon}{
a vector specifying the longitudes of the polygon vertices.
}
  \item{lat}{
a vector specifying the latitudes of the polygon vertices.
}
}
\details{
To be provided ...
}
\value{
To be provided ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Helge Goessling
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{sl.checkposition}}
}
\examples{
## To be provided ...
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
