\name{sl.spatialfilter}
\alias{sl.spatialfilter}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Apply Spatial Filter
}
\description{
Apply a spatial filter to a numerical field.
}
\usage{
sl.spatialfilter(num, sf.gw.res)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{num}{
a vector of length \code{N} specifying the unfiltered numerical field.
}
  \item{sf.gw.res}{
a list (as returned by \code{sl.spatialfilter.getweights}) with the following three elements:
\code{src_address}:
an integer vector of length \code{M} giving the indices of the source points.
\code{dst_address}:
an integer vector of length \code{M} giving the indices of the destination points corresponding to the source points.
\code{filter_weights}:
an integer vector of length \code{M} giving the filter weights.
}
}
\details{
Some details are described in \code{sl.spatialfilter.getweights}.
}
\value{
A vector of length \code{N} giving the filtered numerical field.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Helge Goessling
}
\note{
The filter information generated by \code{sl.spatialfilter.getweights} has (not accidentally) the same format as used by the Climate Data Operators (\code{CDO}) and can be written to a NetCDF file that can be used with \code{CDO} to filter large amounts of data much faster than with spheRlab (using this function).
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{sl.spatialfilter.getweights}
}
\examples{
## To be provided ...
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
