\name{spheRlab-package}
\alias{spheRlab-package}
\alias{spheRlab}
\docType{package}
\title{
\packageTitle{spheRlab}
}
\description{
\packageDescription{spheRlab}
}
\details{
\packageDESCRIPTION{spheRlab}
\packageIndices{spheRlab}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{spheRlab}

Maintainer: \packageMaintainer{spheRlab}
}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
%~~ directory ~~
%\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
\examples{
%
}
