\name{sl.plot.init}
\alias{sl.plot.init}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Initialise spheRlab Plot
}
\description{
Initialise a standard spheRlab plot. Returns a plot specifics list that needs to be provided to any spheRlab plotting functions called afterwards, ended by a call of \code{\link{sl.plot.end}}.
}
\usage{
sl.plot.init(projection = "lonlat", lonlat.lonrange = c(-180, 180), lonlat.latrange = c(-90, 90), polar.lonlatrot = c(0, 90, 0), polar.latbound = 0, regpoly.lonlatrot = c(0, 90, 0), regpoly.N = 3, regpoly.lat0 = 60, regpoly.rotfrac = 0, col.background = NULL, precision = 1, main = "", xshift = 0, yshift = 0, device = "pdf", do.init = TRUE, do.init.device = do.init, file.name = paste0("~/sl.plot.",device), width = 12)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{projection}{
a character specifing the plot projection type. Currently possible projections are \code{"polar"} (orthographic), \code{"lonlat"}, and \code{"regpoly"} (gnomonic) (to be extended).
}
  \item{lonlat.lonrange}{
a vector of length 2 specifying the plot longitude range in ascending order. Only used if \code{projection="lonlat"}.
}
  \item{lonlat.latrange}{
a vector of length 2 specifying the plot latitude range in ascending order. Absolute values above 89 degrees are forbidden. Only used if \code{projection="lonlat"}.
}
  \item{polar.lonlatrot}{
a vector of length 3 specifying the projection North Pole and rotation. See the documentation of \code{\link{sl.lonlatrot2abg}} for details. Only used if \code{projection="polar"}.
}
  \item{polar.latbound}{
a scalar in the range [0,90) specifying the latitude (post-rotation) at which the plot shall be truncated. Default is \code{polar.latbound=0}, in which case the whole hemisphere is plotted.  Only used if \code{projection="polar"}.
}
  \item{regpoly.lonlatrot}{
a vector of length 3 specifying the projection North Pole and rotation. See the documentation of \code{\link{sl.lonlatrot2abg}} for details. Only used if \code{projection="regpoly"}.
}
  \item{regpoly.N}{
an integer specifying the number of vertices of the regular polygon.
}
  \item{regpoly.lat0}{
a scalar in the range (0,90) specifying the latitude (post-rotation) at which the vertices of the regular polygon shall be placed. Default is \code{regpoly.latbound=60} (which does not correspond to a platonian body). Only used if \code{projection="regpoly"}.
}
  \item{regpoly.rotfrac}{
a scalar in the range [0,1) specifying at which longitude (post-rotation) the 'first' vertex of the regular polygon shall be placed. The argument is not the longitude itself, but the fraction by which the vertex shall be rotated from \code{0 degrees} to \code{360/regpoly.N degrees}. Default is \code{regpoly.rotfrac=0}.
}
  \item{col.background}{
a background colour. Default is \code{col.background=NULL}, meaning that no background is drawn. Note that this parameter determines only the backgound colour of the map. To set the background colour of the whole plot, use \code{par(bg=...)} before calling \code{sl.plot.init}.
}
  \item{precision}{
a scalar specifying for projections with curved boundaries the precision (in degrees) at which the backgound polygon boundary is broken into straight line segments. Used only if \code{col.background != NULL}.
}
  \item{main}{
a character specifying the main plot title.
}
  \item{xshift}{
a scalar specifying how far the plot region shall be shifted in x-direction. Default is \code{xshift=0}. Needed for multi-panel plots as generated by \code{\link{sl.plot.platon.init}} and \code{\link{sl.plot.3D.init}}.
}
  \item{yshift}{
a scalar specifying how far the plot region shall be shifted in y-direction. Default is \code{yshift=0}. Needed for multi-panel plots as generated by \code{\link{sl.plot.platon.init}} and \code{\link{sl.plot.3D.init}}.
}
  \item{device}{
a character specifying which graphics device to use. Default is \code{device='pdf'}. Ignored if \code{do.init.device=FALSE}.
}
  \item{do.init}{
a logical value indicating whether or not an initial call of \code{plot} shall be executed. Default is \code{do.init=TRUE}. When called by \code{\link{sl.plot.platon.init}} or \code{\link{sl.plot.3D.init}}, this argument is set \code{FALSE} as only one initial call of \code{plot} is wanted.
}
  \item{do.init.device}{
a logical value indicating whether or not initialise a device as specified by \code{device}. Default is \code{do.init.device=do.init}. If \code{do.init.device=FALSE} but \code{do.init=TRUE}, the plot will be initialised in the standard graphical device.
}
  \item{file.name}{
a character specifying the output file. Only used if \code{do.init.device=TRUE}. Default is \code{file.name=paste0('~/sl.plot.',device)}.
}
  \item{width}{
a scalar giving the width of the resulting figure. The height will be derived according to the chosen projection and other relevant plotting parameters. Only used if \code{do.init.device=TRUE}. Default is \code{width=12}.
}
}
\details{
For a standard single-panel spheRlab plot, this function needs to be called explicitly to initialise the plot, and to generate a plot specifics list that needs to be provided to any spheRlab plotting functions called afterwards to transfer information of the projection, rotation, plot range, and so forth. For a multi-panel plot as generated by \code{\link{sl.plot.platon.init}} or \code{\link{sl.plot.3D.init}}, this function is called internally once for every panel.
}
\value{
A (single-panel) spheRlab plot specifics list.

The first element, always returned, is:
  \item{projection}{
a character specifying the projection.
  }
The last four elements, also always returned, are:
  \item{xlim}{
a vector of length 2 specifying the projected x-range (used also as argument \code{xlim} when calling \code{plot}.
  }
  \item{ylim}{
a vector of length 2 specifying the projected y-range (used also as argument \code{ylim} when calling \code{plot}.
  }
  \item{xshift}{
see the corresponding argument.
}
  \item{yshift}{
see the corresponding argument.
}
Additional list elements inbetween depend on the projection.
For \code{projection="lonlat"}:
  \item{lonlat.lonrange}{
see the corresponding argument.
}
  \item{lonlat.latrange}{
see the corresponding argument.
}
For \code{projection="polar"}:
  \item{polar.lonlatrot}{
see the corresponding argument.
}
  \item{polar.latbound}{
see the corresponding argument.
}
  \item{alpha}{
a scalar specifying the alpha parameter derived from \code{polar.lonlatrot}.
}
  \item{beta}{
a scalar specifying the beta parameter derived from \code{polar.lonlatrot}.
}
  \item{gamma}{
a scalar specifying the gamma parameter derived from \code{polar.lonlatrot}.
}
For \code{projection="regpoly"}:
  \item{regpoly.lonlatrot}{
see the corresponding argument.
}
  \item{regpoly.N}{
see the corresponding argument.
}
  \item{regpoly.lat0}{
see the corresponding argument.
}
  \item{alpha}{
a scalar specifying the alpha parameter derived from \code{regpoly.lonlatrot}.
}
  \item{beta}{
a scalar specifying the beta parameter derived from \code{regpoly.lonlatrot}.
}
  \item{gamma}{
a scalar specifying the gamma parameter derived from \code{regpoly.lonlatrot}.
}
  \item{regpoly.z0}{
a scalar specifying the z-coordinate (post-rotation) of the regular polygon vertices (on a unit sphere).
}
  \item{regpoly.rotfrac}{
see the corresponding argument.
}
  \item{regpoly.cornerlons0}{
a vector of length \code{regpoly.N} specifying the longitudes (post-rotation) of the regular polygon vertices.
}
  \item{regpoly.lat0i}{
a scalar specifying the maximum latitude (post-rotation) of the regular polygon edges (located at the great-circle edge midpoints).
}
  \item{regpoly.cornerlons}{
a vector of length \code{regpoly.N} specifying the longitudes (pre-rotation, that is, in original coordinates) of the regular polygon vertices.
}
  \item{regpoly.cornerlats}{
a vector of length \code{regpoly.N} specifying the latitudes (pre-rotation, that is, in original coordinates) of the regular polygon vertices.
}
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Helge Goessling
}
\note{
Projections to be added include: \code{"robinson"} and \code{"mercator"}.

The projection \code{"regpoly"} corresponds to a polar orthographic projection, truncated along a regular polygon, which allows composing the whole spherical surface with precisely connecting polygons, forming a platonian-type body (e.g. cube or icosahedron). \code{\link{sl.plot.platon.init}} calls this function multiple times to plot onto platonian-type bodies.

Typical multi-panel plotting (beyond the rather special functions \code{\link{sl.plot.platon.init}} or \code{\link{sl.plot.3D.init}}) shall be implemented at some point, too, making use of this function in a similar way as \code{\link{sl.plot.platon.init}} and \code{\link{sl.plot.3D.init}}, but with explicit calls inbetween where the current panel to be plotted into will need to be specified.
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{sl.plot.end}}, \code{\link{sl.plot.platon.init}}, \code{\link{sl.plot.3D.init}}
}
\examples{
## To be provided ...
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
